+++ src/Controler/register.js
@@ -1,8 +1,5 @@
 
-import "./firebase.js";
-import { ponUp } from "./google.js";
-import { logInFn, signUpFn } from './account.js';
-import { facebook } from './facebook.js';
+import { facebook, ponUp, logInFn, signUpFn, logOutFn } from "../Model/firebase.js";
 
 
 document.querySelector('#googleAuth').addEventListener("click", () => {
