+++ src/index.html
@@ -16,22 +16,7 @@
 </head>
 
 <body>
-    <div id="createAccount">
-        <h1>App Feminista</h1>
-        <h2>Crea tu cuenta</h2>
-        <form id="form">
-            <!--input type="text" name="userName" id="registerName" placeholder="nombre">-->
-            <input type="email" name="email" id="emailField" placeholder="correo"></br>
-            <input type="password" name="password" id="passwordField" placeholder="contraseña"></br>
-            <button type="submit" id="logIn">ingresa</button></br>
-            <button type="submit" id="register">regístrate</button></br>
-        </form>
-        <p>Ó inicia sesión con</p>
-        <!--input type="image" src="./images/facebook.png"id="facebook"-->
-        <img src="./images/facebook.png" alt="facebook" id="facebookAuth" class="socialMedia">
-        <img src="./images/google.png" alt="Google" id="googleAuth" class="socialMedia">
-        <img src="./images/twitter (1).png" alt="Twitter" id="twitterAuth" class="socialMedia">
-    </div>
+    <div id="app"></div>
     <div id="welcomeUser">
         <h1>Welcome User</h1>
         <div id="userID"></div>
